Rework length
